"Fixing" the plugin
I'm not even sure if this is the "right way" to do it, but in my case the plugin didn't work at all. I found the issue #120 that solved the problem, but I'm not seeing any pull requests about it.

My code was this example http://gionkunz.github.io/chartist-js/plugins.html#tooltip-plugin and the tooltip didn't show up.